requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django==1.6.2

django-model-utils==2.0

psycopg2==2.5.2
South==0.8.4

djangorestframework==2.3.13
jsonfield==0.9.20
django-uuidfield==0.4.0
drf-nested-routers==0.1.3
django-pagedown==0.0.5
Markdown==2.4
bleach==1.4
django-oauth2-provider==0.2.6

git+ssh://git@github.com/ministryofjustice/cla_common.git@0.1.6#egg=cla_common==0.1.6
